Move now deprecated and unused types to their own kt file
To reduce the noise when reading the actually used code
